Initial dump
dumped core files from SoopremeBot and added a few changes to make this easier to use. It is currently in an unusable state as many problems need to be fixed and a large number of errors arent being handled correctly. This will be fixed in the next commit I (eobsite1) push